Refactor/kafka spring

+++ src/test/java/com/flightmanager/domain/entities/AirportTest.java
@@ -1,7 +1,6 @@
 package com.flightmanager.domain.entities;
 
 import org.junit.Test;
-
 import static org.junit.Assert.assertEquals;
 
 public class AirportTest {
@@ -11,7 +10,7 @@ public class AirportTest {
 	public void testInstantiation() {
 		assertEquals("SBGR", this.airport.getICAO());
 		assertEquals("GRU", this.airport.getIATA());
-		assertEquals(200, this.airport.getGatesAmount());
+		assertEquals(200, (int) this.airport.getGatesAmount());
 		assertEquals(200, this.airport.getGatesList().size());
 	}
 
@@ -21,7 +20,7 @@ public class AirportTest {
 		this.airport.getGate(10).openBoarding();
 		this.airport.updateGateDocking(this.airport.getGate(5), false);
 
-		assertEquals(145, this.airport.getGate(145).getGateNumber());
+		assertEquals(145, (int) this.airport.getGate(145).getGateNumber());
 		assertEquals(false, this.airport.getGate(145).isOpenToBoarding());
 		assertEquals(true, this.airport.getGate(10).isOpenToBoarding());
 		assertEquals(false, this.airport.getGate(5).isFreeToDock());
